+++ src/main/ruby/truffleruby/core/truffle/polyglot.rb
@@ -414,18 +414,21 @@ module Java
   end
 
   def self.import(name)
+    nesting = Primitive.caller_nesting
+    mod = nesting.first || Object
+
     name = name.to_s
     simple_name = name.split('.').last
     type = Java.type(name)
-    if Object.const_defined?(simple_name)
-      current = Object.const_get(simple_name)
+    if mod.const_defined?(simple_name)
+      current = mod.const_get(simple_name)
       if current.equal?(type)
         # Ignore - it's already set
       else
         raise NameError, "constant #{simple_name} already set"
       end
     else
-      Object.const_set simple_name, type
+      mod.const_set simple_name, type
     end
     type
   end